iceberg: moved table_identifier to separate file
Separated table identifier from catalog interface definition.
